Created Archive component to display single Archives before downloading
